third_party/WebKit/LayoutTests/fast/canvas-api/canvas-hit-regions-clear-test.html

<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<body>
<canvas id='canvas' width='400' height='400'>
  <button id='yellow'></button>
</canvas>
<script src='./resources/canvas-hit-region-event.js'></script>
<script>

test(function(t) {
    var canvas = document.getElementById('canvas');
    var context = canvas.getContext('2d');

    assert_equals(internals.countHitRegions(context), 0);

    context.fillStyle = 'red';
    context.rect(0, 0, 100, 100);
    context.fill();
    context.addHitRegion({ id : 'red' });
    assert_equals(internals.countHitRegions(context), 1);

    context.beginPath();
    context.fillStyle = 'yellow';
    context.rect(100, 100, 50, 50);
    context.fill();
    context.addHitRegion({ id : 'yellow', control : document.getElementById('yellow') });
    assert_equals(internals.countHitRegions(context), 2);

    assert_equals(clickCanvas(60, 60), 'red');
    assert_equals(clickCanvas(120, 120), 'yellow');

    context.clearRect(50, 50, 50, 50);
    assert_equals(internals.countHitRegions(context), 2);
    assert_equals(clickCanvas(60, 60), null);
    assert_equals(clickCanvas(120, 120), 'yellow');

    context.clearRect(100, 100, 50, 50);
    assert_equals(internals.countHitRegions(context), 1);
    assert_equals(clickCanvas(120, 120), null);

    context.beginPath();
    context.fillStyle = 'red';
    context.rect(60, 60, 40, 40);
    context.fill();
    assert_equals(clickCanvas(40, 40), 'red');

    context.addHitRegion({ id : 'red' });
    assert_equals(internals.countHitRegions(context), 1);
    assert_equals(clickCanvas(62, 62), 'red');
    assert_equals(clickCanvas(40, 40), null);

    context.beginPath();
    context.fillStyle = 'yellow';
    context.rect(0, 0, 50, 50);
    context.fill();
    context.addHitRegion({ id : 'yellow', control : document.getElementById('yellow') });
    assert_equals(internals.countHitRegions(context), 2);
    assert_equals(clickCanvas(40, 40), 'yellow');

    context.beginPath();
    context.fillStyle = 'yellow';
    context.rect(100, 0, 50, 50);
    context.fill();
    context.addHitRegion({ control : document.getElementById('yellow') });
    assert_equals(internals.countHitRegions(context), 2);
    assert_equals(clickCanvas(40, 40), null);
    assert_equals(clickCanvas(101, 1), null);

    context.beginPath();
    context.fillStyle = 'blue';
    context.rect(100, 50, 20, 20);
    context.fill();
    context.addHitRegion({ id : 'blue' });
    assert_equals(internals.countHitRegions(context), 3);
    assert_equals(clickCanvas(101, 51), 'blue');

    context.removeHitRegion('blue');
    assert_equals(internals.countHitRegions(context), 2);
    assert_equals(clickCanvas(101, 51), null);

    context.clearHitRegions();
    assert_equals(internals.countHitRegions(context), 0);
    assert_equals(clickCanvas(62, 62), null);

    context.clearRect(0, 0, 400, 400);
    assert_equals(internals.countHitRegions(context), 0);

}, 'Test canvas hit regions: clear test.');

</script>
</body>
